feat : update 24 07 23

+++ FamTec/FamTec/FamTec/Shared/Client/DTO/Normal/Users/UserPermDTO.cs
@@ -10,31 +10,31 @@ namespace FamTec.Shared.Client.DTO.Normal.Users
     public class UserPermDTO
     {
         [Display(Name ="기본 정보")]
-        public sbyte BasicPerm { get; set; }
+        public int BasicPerm { get; set; }
         [Display(Name = "기계 설비")]
-        public sbyte MachinPerm { get; set; }
+        public int MachinPerm { get; set; }
         [Display(Name = "전기 설비")]
-        public sbyte ElecPerm { get; set; }
+        public int ElecPerm { get; set; }
         [Display(Name = "승강 설비")]
-        public sbyte LiftPerm { get; set; }
+        public int LiftPerm { get; set; }
         [Display(Name = "소방 설비")]
-        public sbyte FirePerm { get; set; }
+        public int FirePerm { get; set; }
         [Display(Name = "건축 설비")]
-        public sbyte ConstructPerm { get; set; }
+        public int ConstructPerm { get; set; }
         [Display(Name = "통신 설비")]
-        public sbyte NetworkPerm { get; set; }
+        public int NetworkPerm { get; set; }
         [Display(Name = "미화 설비")]
-        public sbyte BeautyPerm { get; set; }
+        public int BeautyPerm { get; set; }
         [Display(Name = "보안 설비")]
-        public sbyte SecurityPerm { get; set; }
+        public int SecurityPerm { get; set; }
         [Display(Name = "자재 관리")]
-        public sbyte MaterialPerm { get; set; }
+        public int MaterialPerm { get; set; }
         [Display(Name = "에너지 관리")]
-        public sbyte EnergyPerm { get; set; }
+        public int EnergyPerm { get; set; }
         [Display(Name = "사용자 관리")]
-        public sbyte UserPerm { get; set; }
+        public int UserPerm { get; set; }
         [Display(Name = "민원 관리")]
-        public sbyte VocPerm { get; set; }
+        public int VocPerm { get; set; }
 
 
     }
